working on Quarto; try to execute and get output.
working on Quarto; try to execute and get output.

working on Quarto; try to execute and get output.

new build of vocab docs

working on Quarto; try to execute and get output.

working on Quarto; try to execute and get output.

working on Quarto; try to execute and get output.